requirements.txt
pyparsing==3.0.9
PyYAML==6.0
kubernetes==25.3.0
